refactor(components): migrate ModalWrapper to TypeScript
